2010-02-14  Noel O'Boyle
	* test/*: Avoid SafeOpen warnings by including obutil.h
	* src/stereo/perception.cpp: Added default to switch statement to
	avoid compiler warning about "not all paths return a value"
	* src/formats/gausscubeformat.cpp, cifformat.cpp: Added some static
	casts to avoid warnings
	* src/formats/gaussformat.cpp: Added explicit comparison against NULL
	to avoid warning
	* src/formats/cifformat.cpp: Disabled warning 4503 about very long
	decorated name being truncated.

+++ src/formats/gausscubeformat.cpp
@@ -574,7 +574,7 @@ bool OBGaussianCubeFormat::ReadMolecule( OBBase* pOb, OBConversion* pConv )
     gd->GetOriginVector(origin);
 
     // line 3: number of atoms, origin x y z
-    snprintf(buffer, BUFF_SIZE,"%5d%12.6f%12.6f%12.6f", -mol.NumAtoms(),
+    snprintf(buffer, BUFF_SIZE,"%5d%12.6f%12.6f%12.6f", - static_cast<signed int> (mol.NumAtoms()),
         origin[0]*ANGSTROM_TO_BOHR, origin[1]*ANGSTROM_TO_BOHR, origin[2]*ANGSTROM_TO_BOHR);
     ofs << buffer << endl;
